+++ packages/components/atoms/f-filter-pill/src/components/FilterPillSkeleton.vue
